Moves AuthenticationTokenResolver and splits out interface
