+++ src/Altinn.App.Core/Internal/Auth/IAuthenticationTokenResolver.cs
@@ -1,0 +1,18 @@
+using Altinn.App.Core.Models;
+using Altinn.App.Core.Models.AuthenticationMethod;
+
+namespace Altinn.App.Core.Internal.Auth;
+
+/// <summary>
+/// Resolves authentication tokens based on the specified authentication method.
+/// </summary>
+internal interface IAuthenticationTokenResolver
+{
+    /// <summary>
+    /// Retrieves an access token based on the specified authentication method.
+    /// </summary>
+    Task<JwtToken> GetAccessToken(
+        AuthenticationMethod authenticationMethod,
+        CancellationToken cancellationToken = default
+    );
+}
